Modified pre-install script to be more 'sh' friendly.

+++ scripts/pre-install.sh
@@ -1,6 +1,4 @@
 #!/bin/sh
 
 # Copy existing configuration if pre-existing installation is found
-if -f /etc/opt/influxdb/influxdb.conf ; then
-    cp -a /etc/opt/influxdb /etc/influxdb
-fi
+test -f /etc/opt/influxdb/influxdb.conf && cp -a /etc/opt/influxdb /etc/influxdb
